+++ composeApp/src/commonMain/kotlin/com/dna/payments/kmm/presentation/ui/features/online_payments/charge/OnlinePaymentChargeViewModel.kt
@@ -66,7 +66,7 @@ class OnlinePaymentChargeViewModel(
         screenModelScope.launch {
             val result = chargePaymentOperationUseCase(
                 transactionId = transactionId,
-                amount = currentState.amount.input.value.toInt()
+                amount = currentState.amount.input.value.toDouble()
             )
             setState {
                 copy(
